pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repro-mcp"
version = "0.1.0"
description = "Reproducibility logging and rule enforcement MCP server for scientific computing"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
repro-mcp = "repro_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/repro_mcp"]
